refactor(azure-storage): vendor non-text extension map, mirror sync/async unit tests per method
Address the latest Deep Agents backend review round:

- Vendor the non-text extension set (from deepagents 0.6.12's private
  _EXTENSION_TO_FILE_TYPE) into _utils.is_text_file() so the backend no
  longer imports the private _get_file_type helper; a parity test fails in
  CI if the vendored copy drifts from the installed deepagents.
- Split the monolithic unit test file into per-method modules under
  tests/unit_tests/deepagents/backend/, mirroring every case across the
  sync and async forks and dissolving the edge-case grab-bag classes into
  their method classes (path-utility cases move to test_utils.py).
- Assert client call wiring (blob key resolution, upload/download/listing
  kwargs) in each method's basic call-pattern test; async_list now wraps
  its generator in a MagicMock so listing call args are recordable.
- Tighten the sync round-trip integration assertions to exact matches.
