+++ libs/azure-storage/tests/unit_tests/deepagents/test_utils.py
@@ -9,9 +9,11 @@ import pytest
 pytest.importorskip("deepagents")
 
 from langchain_azure_storage.deepagents._utils import (  # noqa: E402
+    _NON_TEXT_EXTENSIONS,
     build_file_info,
     from_blob_key,
     get_prefix_for_path,
+    is_text_file,
     normalize_path,
     to_blob_key,
 )
@@ -23,6 +25,12 @@ class TestNormalizePath:
 
     def test_root(self) -> None:
         assert normalize_path("/") == ""
+
+    def test_empty_string(self) -> None:
+        assert normalize_path("") == ""
+
+    def test_trailing_slash(self) -> None:
+        assert normalize_path("/src/") == "src"
 
     def test_double_slashes(self) -> None:
         assert normalize_path("//src//main.py") == "src/main.py"
@@ -39,6 +47,9 @@ class TestNormalizePath:
 class TestToBlobKey:
     def test_with_prefix(self) -> None:
         assert to_blob_key("workspace/", "/src/main.py") == "workspace/src/main.py"
+
+    def test_root(self) -> None:
+        assert to_blob_key("workspace/", "/") == "workspace/"
 
     def test_without_prefix(self) -> None:
         assert to_blob_key("", "/src/main.py") == "src/main.py"
@@ -65,6 +76,9 @@ class TestGetPrefixForPath:
     def test_subdir_no_prefix(self) -> None:
         assert get_prefix_for_path("", "/src") == "src/"
 
+    def test_root_no_prefix(self) -> None:
+        assert get_prefix_for_path("", "/") == ""
+
 
 class TestBuildFileInfo:
     def test_defaults(self) -> None:
@@ -77,3 +91,35 @@ class TestBuildFileInfo:
 
     def test_directory(self) -> None:
         assert build_file_info("/src/", is_dir=True)["is_dir"] is True
+
+
+class TestIsTextFile:
+    def test_text_extension(self) -> None:
+        assert is_text_file("/src/main.py") is True
+
+    def test_non_text_extension(self) -> None:
+        assert is_text_file("/img/logo.png") is False
+
+    def test_extension_is_case_insensitive(self) -> None:
+        assert is_text_file("/img/LOGO.PNG") is False
+
+    def test_no_extension_defaults_to_text(self) -> None:
+        assert is_text_file("/notes") is True
+
+    def test_unknown_extension_defaults_to_text(self) -> None:
+        assert is_text_file("/data.custom") is True
+
+    def test_vendored_set_matches_installed_deepagents(self) -> None:
+        # Drift canary for the vendored `_NON_TEXT_EXTENSIONS` set: it must
+        # classify exactly like the installed deepagents' private
+        # `_get_file_type` helper. If a deepagents release adds, removes, or
+        # relocates extensions, this fails in CI instead of the backend
+        # silently classifying reads differently from the reference backends.
+        from deepagents.backends.utils import _EXTENSION_TO_FILE_TYPE, _get_file_type
+
+        upstream_non_text = {
+            ext
+            for ext in _EXTENSION_TO_FILE_TYPE
+            if _get_file_type(f"f{ext}") != "text"
+        }
+        assert upstream_non_text == _NON_TEXT_EXTENSIONS
